python/requirements.txt
fastapi
uvicorn[standard]
pydantic
pandas
numpy
scikit-learn
xgboost
statsmodels
python-dotenv
httpx
